requirements.txt
streamlit==1.40.2
coolname==2.2.0
streamlit_javascript==0.1.5
PyGithub==2.5.0
